+++ src/components/MusicApp/navbar/NavBar.js
