+++ src/components/MusicApp/navbar/SearchBar.js
